Moved class creation to queryableperson, moved files around.
