+++ HousingSearchApi.Tests/V1/Helper/SearchPersonsQueryContainerOrchestratorTests.cs
@@ -2,6 +2,7 @@ using System.Linq;
 using FluentAssertions;
 using HousingSearchApi.V1.Boundary.Requests;
 using HousingSearchApi.V1.Domain.ElasticSearch;
+using HousingSearchApi.V1.Factories;
 using HousingSearchApi.V1.Interfaces;
 using Nest;
 using Xunit;
