+++ HousingSearchApi.Tests/V1/Helper/TestDataHelper.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using AutoFixture;
 using HousingSearchApi.V1.Domain.ElasticSearch;
+using HousingSearchApi.V1.Factories;
 using Nest;
 
 namespace HousingSearchApi.Tests.V1.Helper
