+++ HousingSearchApi/HousingSearchApi.csproj
@@ -33,8 +33,4 @@
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.3" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="V1\Factories\" />
-  </ItemGroup>
-
 </Project>
