+++ HousingSearchApi/V1/Domain/Person.cs
@@ -1,71 +1,9 @@
 using System.Collections.Generic;
-using HousingSearchApi.V1.Domain.ElasticSearch;
 
 namespace HousingSearchApi.V1.Domain
 {
     public class Person
     {
-        public static Person Create(QueryablePerson person)
-        {
-            return new Person
-            {
-                Id = person.Id,
-                Title = person.Title,
-                Firstname = person.Firstname,
-                MiddleName = person.MiddleName,
-                Surname = person.Surname,
-                PreferredFirstname = person.PreferredFirstname,
-                PreferredSurname = person.PreferredSurname,
-                Ethinicity = person.Ethinicity,
-                Nationality = person.Nationality,
-                PlaceOfBirth = person.PlaceOfBirth,
-                DateOfBirth = person.DateOfBirth,
-                Gender = person.Gender,
-                Identification = Create(person.Identification ?? new List<ElasticSearch.Identification>()),
-                PersonTypes = person.PersonTypes,
-                IsPersonCautionaryAlert = person.IsPersonCautionaryAlert,
-                IsTenureCautionaryAlert = person.IsTenureCautionaryAlert,
-                Tenures = Create(person.Tenures ?? new List<ElasticSearch.Tenures>())
-            };
-        }
-
-        private static List<Identification> Create(List<ElasticSearch.Identification> identifications)
-        {
-            var identList = new List<Identification>();
-
-            foreach (ElasticSearch.Identification identification in identifications)
-            {
-                identList.Add(new Identification
-                {
-                    IdentificationType = identification.IdentificationType,
-                    LinkToDocument = identification.LinkToDocument,
-                    OriginalDocumentSeen = identification.OriginalDocumentSeen,
-                    Value = identification.Value
-                });
-            }
-
-            return identList;
-        }
-
-        private static List<Tenures> Create(List<ElasticSearch.Tenures> tenures)
-        {
-            var tenureList = new List<Tenures>();
-
-            foreach (ElasticSearch.Tenures tenure in tenures)
-            {
-                tenureList.Add(new Tenures
-                {
-                    AssetFullAddress = tenure.AssetFullAddress,
-                    EndDate = tenure.EndDate,
-                    Id = tenure.Id,
-                    StartDate = tenure.StartDate,
-                    Type = tenure.Type
-                });
-            }
-
-            return tenureList;
-        }
-
         public string Id { get; set; }
 
         public string Title { get; set; }
