+++ HousingSearchApi/V1/Interfaces/ISearchPersonElasticSearchHelper.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using HousingSearchApi.V1.Boundary.Requests;
 using HousingSearchApi.V1.Domain.ElasticSearch;
+using HousingSearchApi.V1.Factories;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces
