+++ HousingSearchApi/V1/Interfaces/SearchPersonsQueryContainerOrchestrator.cs
@@ -1,5 +1,6 @@
 using HousingSearchApi.V1.Boundary.Requests;
 using HousingSearchApi.V1.Domain.ElasticSearch;
+using HousingSearchApi.V1.Factories;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces
