+++ HousingSearchApi/V1/Interfaces/Sorting/IPersonListSort.cs
@@ -1,4 +1,5 @@
 using HousingSearchApi.V1.Domain.ElasticSearch;
+using HousingSearchApi.V1.Factories;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces.Sorting
